Move list of default views to own file
