+++ crates/viewer/re_viewer/src/app.rs
@@ -220,13 +220,11 @@ impl App {
             state.app_options.video_decoder_hw_acceleration = video_decoder_hw_acceleration;
         }
 
-        let mut view_class_registry = ViewClassRegistry::default();
-        if let Err(err) = populate_view_class_registry_with_builtin(&mut view_class_registry) {
-            re_log::error!(
-                "Failed to populate the view type registry with built-in views: {}",
-                err
-            );
-        }
+        let view_class_registry = crate::default_views::create_view_class_registry()
+            .unwrap_or_else(|err| {
+                re_log::error!("Failed to create view class registry: {err}");
+                Default::default()
+            });
 
         #[allow(unused_mut, clippy::needless_update)] // false positive on web
         let mut screenshotter = crate::screenshotter::Screenshotter::default();
@@ -2459,26 +2457,6 @@ impl eframe::App for App {
     fn as_any_mut(&mut self) -> Option<&mut dyn std::any::Any> {
         Some(&mut *self)
     }
-}
-
-/// Add built-in views to the registry.
-fn populate_view_class_registry_with_builtin(
-    view_class_registry: &mut ViewClassRegistry,
-) -> Result<(), ViewClassRegistryError> {
-    re_tracing::profile_function!();
-    view_class_registry.add_class::<re_view_bar_chart::BarChartView>()?;
-    view_class_registry.add_class::<re_view_dataframe::DataframeView>()?;
-    view_class_registry.add_class::<re_view_graph::GraphView>()?;
-    #[cfg(feature = "map_view")]
-    view_class_registry.add_class::<re_view_map::MapView>()?;
-    view_class_registry.add_class::<re_view_spatial::SpatialView2D>()?;
-    view_class_registry.add_class::<re_view_spatial::SpatialView3D>()?;
-    view_class_registry.add_class::<re_view_tensor::TensorView>()?;
-    view_class_registry.add_class::<re_view_text_document::TextDocumentView>()?;
-    view_class_registry.add_class::<re_view_text_log::TextView>()?;
-    view_class_registry.add_class::<re_view_time_series::TimeSeriesView>()?;
-
-    Ok(())
 }
 
 fn paint_background_fill(ui: &egui::Ui) {
