+++ crates/viewer/re_viewer/src/lib.rs
@@ -7,6 +7,7 @@ mod app;
 mod app_blueprint;
 mod app_state;
 mod background_tasks;
+mod default_views;
 pub mod env_vars;
 pub mod event;
 mod navigation;
